repurposed animation service as a animation component factory
